feat: enhance delete contact functionality with error handling; re-enable React.StrictMode in main entry

+++ frontend/src/components/table-components/DeleteButton.tsx
@@ -19,7 +19,11 @@ const DeleteButton: FC<{ contact: Partial<IContact>; listing_id: string }> = pro
   const deleteMutation = useMutation({
     mutationKey: ['delete_contact', _id],
     mutationFn: () => {
-      return deleteContact(listing_id, _id || '')
+      if (deleteContact) {
+        return deleteContact(listing_id, _id || '')
+      } else {
+        return Promise.reject('')
+      }
     },
     onSuccess() {
       toast({ title: 'Contact Deleted', description: `Deleted ${number}` })
